begin moving over host code
